pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mixture-design-doe"
version = "0.1.0"
description = "High-efficiency mixture design of experiments"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "numpy>=1.20",
    "pandas>=1.3",
    "matplotlib>=3.3",
    "seaborn>=0.11",
    "scipy>=1.7",
    "scikit-learn>=0.24",
    "streamlit>=1.0",
    "openpyxl>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.9",
]
viz = ["python-ternary>=1.0"]

[project.urls]
Homepage = "https://github.com/yourusername/mixture-design-doe"
Documentation = "https://mixture-design-doe.readthedocs.io"
Repository = "https://github.com/yourusername/mixture-design-doe"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/examples/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
